feat(action-tool-installer): add shared binary-fetching package
Extract the generic download/extract/cache engine from setup-k8s-tools
into a new internal @internal/action-tool-installer workspace package so
future CLI installers (e.g. oidc-token-cli) reuse it. Parameterize the
cache namespace, refactor setup-k8s-tools to consume it via workspace:^,
and quietly skip caching when the actions cache is read-only for
untrusted triggers (CacheWriteDeniedError).
