+++ packages/action-tool-installer/eslint.config.ts
@@ -1,0 +1,7 @@
+import { defineConfig } from "eslint/config";
+import { base, configFiles } from "@abinnovision/eslint-config-base";
+
+export default defineConfig([
+	{ extends: [base] },
+	{ files: ["*.{c,m,}{t,j}s"], extends: [configFiles] },
+]);
