+++ packages/action-tool-installer/src/index.ts
@@ -1,0 +1,3 @@
+export * from "./install.js";
+export * from "./latest-tag-cache.js";
+export * from "./tools.js";
