+++ packages/action-tool-installer/tsconfig.json
@@ -1,0 +1,10 @@
+{
+	"extends": "../../tsconfig.base.json",
+	"compilerOptions": {
+		"rootDir": "./src",
+		"noEmit": false,
+		"outDir": "./dist",
+		"declaration": true
+	},
+	"include": ["src/**/*.ts"]
+}
